requirements.txt
numpy<2
cupy-cuda12x
